requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pytesseract==0.3.13
pillow==10.4.0
pdf2image==1.17.0
python-docx==1.1.2
